refactor scheduling/shifts routes
