+++ www/app.js
@@ -30,15 +30,7 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-for (var i in routes) {
-    for (var j in routes[i]) {
-        if (j == 'index') {
-            app.use('/'+i, routes[i][j]);
-        } else {
-            app.use('/'+i+'/'+j, routes[i][j]);
-        }
-    }
-}
+app.use('/scheduling', require('./routes/scheduling'))
 
 /// catch 404 and forwarding to error handler
 app.use(function(req, res, next) {
